Rename sampler_v2 module to executor_sampler
