+++ qiskit_ibm_runtime/executor_sampler/post_processors/post_processor_v0_1.py
@@ -31,10 +31,10 @@ def sampler_v2_post_processor_v0_1(result: QuantumProgramResult) -> PrimitiveRes
     """Convert a quantum program result to a primitives result, for a V2 sampler.
 
     Convert :class:`~.QuantumProgramResult` to a :class:`~qiskit.primitives.PrimitiveResult`,
-    for :class:`~qiskit_ibm_runtime.sampler_v2.SamplerV2`.
+    for :class:`~qiskit_ibm_runtime.executor_sampler.SamplerV2`.
 
     This function transforms the raw quantum program execution results into the
-    format expected by :class:`~qiskit_ibm_runtime.sampler_v2.SamplerV2`,
+    format expected by :class:`~qiskit_ibm_runtime.executor_sampler.SamplerV2`,
     creating :class:`~qiskit.primitives.containers.BitArray` objects and
     :class:`~qiskit.primitives.containers.SamplerPubResult` containers for each pub.
 
@@ -42,7 +42,7 @@ def sampler_v2_post_processor_v0_1(result: QuantumProgramResult) -> PrimitiveRes
         result: The raw quantum program result containing measurement data.
 
     Returns:
-        Primitive result for :class:`~qiskit_ibm_runtime.sampler_v2.SamplerV2`.
+        Primitive result for :class:`~qiskit_ibm_runtime.executor_sampler.SamplerV2`.
     """
     if len(result) == 0:
         return PrimitiveResult([])
