+++ test/unit/sampler_v2/post_processors/test_executor_metadata_to_sampler_metadata.py
@@ -20,7 +20,7 @@ from datetime import datetime
 from qiskit_ibm_runtime.execution_span import DoubleSliceSpan, TwirledSliceSpanV2
 from qiskit_ibm_runtime.quantum_program.quantum_program_result import ChunkSpan, Metadata, ChunkPart
 
-from qiskit_ibm_runtime.sampler_v2.post_processors.utils import (
+from qiskit_ibm_runtime.executor_sampler.post_processors.utils import (
     executor_metadata_to_sampler_metadata,
 )
