+++ test/unit/test_runtime_job_post_processing.py
@@ -22,7 +22,7 @@ from qiskit_ibm_runtime.quantum_program.quantum_program_result import (
     QuantumProgramResult,
     Metadata,
 )
-from qiskit_ibm_runtime.sampler_v2 import SAMPLER_POST_PROCESSORS
+from qiskit_ibm_runtime.executor_sampler import SAMPLER_POST_PROCESSORS
 
 
 class TestRuntimeJobPostProcessing(unittest.TestCase):
